requirements.txt
openai>=1.0.0
google-generativeai>=0.3.0
anthropic>=0.8.0
python-dotenv>=1.0.0 